Altered event model to properly display event type
Previously, utilized TYPES nested tuple outside of model declaration, this time it does it inside with variable declaration, allowing us to properly print the event type on both the event information page and the event list (happening now) page.  Also altered to display in descending order based on the event's date and time.